[FLINK-39590] Move DefaultConnectionFactory into the catalog package
Addresses PR review feedback: relocate DefaultConnectionFactory (and its
test) from o.a.f.table.factories to o.a.f.table.catalog so the default
implementation lives next to the connection/catalog types it operates on.
Update the SPI service entry, the ConnectionFactory javadoc @see, and the
CatalogManager import accordingly. Also trim the duplicated sensitive
field list in the class javadoc and point readers to SENSITIVE_FIELD_NAMES
as the single source of truth.
