+++ flink-table/flink-table-common/src/test/java/org/apache/flink/table/catalog/DefaultConnectionFactoryTest.java
@@ -16,11 +16,9 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.factories;
+package org.apache.flink.table.catalog;
 
 import org.apache.flink.table.api.ValidationException;
-import org.apache.flink.table.catalog.CatalogConnection;
-import org.apache.flink.table.catalog.SensitiveConnection;
 import org.apache.flink.table.secret.ReadableSecretStore;
 import org.apache.flink.table.secret.WritableSecretStore;
 import org.apache.flink.table.secret.exceptions.SecretException;
